requirements.txt
aniso8601==2.0.1
appdirs==1.4.3
astroid==1.6.1
attrs==17.4.0
click==6.7
coverage==4.5.1
factory-boy==2.10.0
Faker==0.8.11
Flask==0.12.2
flask-restplus==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.3.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
packaging==16.8
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.2
pylint==1.8.2
pyparsing==2.2.0
pytest==3.4.1
pytest-cov==2.5.1
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-html==1.16.1
pytest-metadata==1.6.0
pytest-sugar==0.9.1
python-dateutil==2.6.0
pytz==2018.3
six==1.11.0
SQLAlchemy==1.2.4
termcolor==1.1.0
text-unidecode==1.2
Werkzeug==0.14.1
wrapt==1.10.11
